backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "pydantic-settings>=2.7.1",
    "pyjwt[crypto]>=2.10.1",
    "passlib[bcrypt]>=1.7.4",
    "aiosmtplib>=3.0.1",
    "httpx>=0.28.1",
]
